requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
docopt==0.6.2
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
pipreqs==0.4.11
py-cord==2.0.0b1
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.8
yarg==0.1.9
yarl==1.7.2
